Update tests to not rely on legacy method cache
Usage of nqp::setmethcache is on the way out; the new dispatcher caches
at callsites and relies on what is returned from `find_method`.

The most common change here is to avoid use of the precomputed method
cache for a method that was never really possible to resolve via the
MOP anyway; this was kind of convenient for setting up test types, but
weird in general.

Some tests go away completely because they were entirely focused on
testing the legacy method cache.