Produce simpler DPs for method dispatches
Prior to this, we always used the `boot-code` dispatcher and added
guards, assuming that the code value we were invoking may change
between dispatches. However, for a method lookup, we resolve it one time
and it's constant. Use that knowledge to avoid adding guards and to
delegate to `boot-code-constant` instead, which permits a range of
further optimizations.